GEODE-5825: Moved tests from integration to unit tests.
