+++ cppcache/integration-test-2/CMakeLists.txt
@@ -31,8 +31,7 @@ add_executable(integration-test-2
   StructTest.cpp
   EnableChunkHandlerThreadTest.cpp
   DataSerializableTest.cpp
-  FunctionExecutionTest.cpp
-  TcrMessageTest.cpp)
+  FunctionExecutionTest.cpp)
 
 target_compile_definitions(integration-test-2
   PUBLIC
@@ -47,7 +46,7 @@ target_include_directories(integration-test-2
 
 target_link_libraries(integration-test-2
   PUBLIC
-    apache-geode-static
+    apache-geode
     testobject
     ACE
     GTest::GTest
